requirements.txt
PyMySQL==1.0.2
PyYAML==6.0
redis==4.3.4
cryptography==2.4.2
requests==2.27.1